fixed numpy 1.12 TypeError

+++ python/render.py
@@ -40,7 +40,7 @@ def vec2im(V, shape = () ):
     '''
 
     if len(shape) < 2:
-        shape = [np.sqrt(V.size)]*2
+        shape = [int(np.sqrt(V.size))]*2
 
     return np.reshape(V, shape)
